Remove QtConcurrent feature
The use of `QtConcurrent::blockingMap` is replaced
with `tbb::parallel_for`.
To note: the setting of the number of threads in tbb
is deprecated. Thus the code now ignores the setting
of this parameter.
The function calls specifying the number of threads
are kept for back-compatibility but the parameter
is ignored.